+++ internal/provider/common.go
@@ -72,7 +72,7 @@ func requireReplaceIfStateContainsPEMString() tfsdk.AttributePlanModifier {
 		stateValue := state.(types.String).Value
 
 		// If the value is indeed a PEM, and
-		if regexp.MustCompile(`^-----BEGIN [[:alpha:]]+-----\n(.|\s)+\n-----END [[:alpha:]]+-----\n$`).MatchString(stateValue) {
+		if regexp.MustCompile(`^-----BEGIN [[:alpha:] ]+-----\n(.|\s)+\n-----END [[:alpha:] ]+-----\n?$`).MatchString(stateValue) {
 			return true, nil
 		}